Uncommented database migration in Initialize method
In the `DataInitializer.cs` file within the `ProjectApi.Data` namespace, the line of code `_context.Database.Migrate();` was uncommented in the `Initialize` method. This change ensures that the database migration is performed when the `Initialize` method is called, keeping the database schema up-to-date with the latest changes in the code.