Update PatchProject method in GitProjectController
The PatchProject method in GitProjectController.cs has been updated. The type of the patchDoc parameter has been changed to JsonPatchDocument<GitProjectDTO>. The retrieval of the gitProject object from the database now includes related Technologies and TechIcon entities. The patchDoc is now applied to a new GitProjectDTO object, and the properties of the gitProject object are updated from this DTO. The Technologies of the gitProject are also updated based on the DTO.