Update CV to reflect confident self-presentation
Revised the "Om mig" section to showcase a more assertive self-description, transitioning from a student to a knowledgeable .NET developer. Updated the "Proffessionella mål" section to indicate an active job search for a Junior .NET developer position, emphasizing a desire to contribute and grow technically. Changed the title in the "Landing Page" from "Studerar till att bli .NET-utvecklare" to "Junior .NET-utvecklare" and updated the CV link to a new file name.